requirements.txt
fastapi
uvicorn
python-docx
azure-storage-blob
python-multipart
PyMuPDF
openai